+++ src/test/java/nl/lexemmens/podman/BuildMojoTest.java
@@ -408,13 +408,13 @@ public class BuildMojoTest extends AbstractMojoTest {
         verify(log, times(1)).info("Detected multistage Containerfile...");
 
         // At random verify some lines
-        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base', matches: true");
-        verify(log, times(1)).debug("Processing line: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null', matches: false");
+        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base'");
+        verify(log, times(1)).debug("Processing candidate: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null'");
 
         // Verify stage detection
-        verify(log, times(1)).debug("Found stage in Containerfile: base");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase2");
+        verify(log, times(1)).debug("Processing stage in Containerfile: base");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase2");
 
         // Verify hashes for stages
         verify(log, times(1)).info("Final image for stage base is: 7e72c870614");
@@ -464,13 +464,73 @@ public class BuildMojoTest extends AbstractMojoTest {
         verify(log, times(1)).info("Detected multistage Containerfile...");
 
         // At random verify some lines
-        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base', matches: true");
-        verify(log, times(1)).debug("Processing line: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null', matches: false");
+        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base'");
+        verify(log, times(1)).debug("Processing candidate: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null'");
 
         // Verify stage detection
-        verify(log, times(1)).debug("Found stage in Containerfile: base");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase2");
+        verify(log, times(1)).debug("Processing stage in Containerfile: base");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase2");
+
+        // Verify hashes for stages
+        verify(log, times(1)).info("Final image for stage base is: 7e72c870614");
+        verify(log, times(1)).info("Final image for stage phase is: 7f55eab001a");
+        verify(log, times(1)).info("Final image for stage phase2 is: d2efc6645cb");
+
+        // Verify tagging image
+        verify(log, times(1)).info("Tagging container image 7f55eab001a from stage phase as registry.example.com/image-name-number-1:0.2.1");
+        verify(log, times(1)).info("Tagging container image d2efc6645cb from stage phase2 as registry.example.com/image-name-number-2:0.2.1");
+        verify(log, times(0)).info("Tagging container image d2efc6645cb as registry.example.com/sample:1.0.0");
+
+        verify(podmanExecutorService, times(1)).tag("7f55eab001a", "registry.example.com/image-name-number-1:0.2.1");
+        verify(podmanExecutorService, times(1)).tag("d2efc6645cb", "registry.example.com/image-name-number-2:0.2.1");
+
+        verify(log, times(1)).info("Built container image.");
+    }
+
+    @Test
+    public void testMultiStageContainerFileWithMultilineOutputStep() throws MojoExecutionException, IOException, URISyntaxException {
+        URI sampleBuildOutputUri = PushMojoTest.class.getResource("/multistagecontainerfile/samplebuildoutput_multiline_step.txt").toURI();
+        Path sampleBuildOutputPath = Paths.get(sampleBuildOutputUri);
+
+        List<String> buildOutputUnderTest = null;
+        try (Stream<String> buildSampleOutput = Files.lines(sampleBuildOutputPath)) {
+            buildOutputUnderTest = buildSampleOutput.collect(Collectors.toList());
+        }
+
+        Assertions.assertNotNull(buildOutputUnderTest);
+
+        PodmanConfiguration podman = new TestPodmanConfigurationBuilder().setTlsVerify(TlsVerify.FALSE).build();
+        ImageConfiguration image = new TestImageConfigurationBuilder("sample")
+                .setContainerfileDir("src/test/resources/multistagecontainerfile")
+                .setTags(new String[]{"0.2.1"})
+                .setCreateLatestTag(false)
+                .setUseCustomImageNameForMultiStageContainerfile(true)
+                .addCustomImageNameForBuildStage("phase", "image-name-number-1")
+                .addCustomImageNameForBuildStage("phase2", "image-name-number-2")
+                .build();
+        configureMojo(podman, image, true, false, false, false, true);
+
+        when(mavenProject.getBuild()).thenReturn(build);
+        when(build.getDirectory()).thenReturn("target");
+        when(serviceHubFactory.createServiceHub(isA(Log.class), isA(MavenProject.class), isA(MavenFileFilter.class), isA(PodmanConfiguration.class), isA(Settings.class), isA(SettingsDecrypter.class))).thenReturn(serviceHub);
+        when(serviceHub.getContainerfileDecorator()).thenReturn(containerfileDecorator);
+        when(serviceHub.getPodmanExecutorService()).thenReturn(podmanExecutorService);
+        when(podmanExecutorService.build(isA(ImageConfiguration.class))).thenReturn(buildOutputUnderTest);
+
+        buildMojo.execute();
+
+        // Verify logging
+        verify(log, times(1)).info("Detected multistage Containerfile...");
+
+        // At random verify some lines
+        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base'");
+        verify(log, times(1)).debug("Processing candidate: 'STEP 8: LABEL Build-User=sample-user2 Git-Repository-Url=null'");
+
+        // Verify stage detection
+        verify(log, times(1)).debug("Processing stage in Containerfile: base");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase2");
 
         // Verify hashes for stages
         verify(log, times(1)).info("Final image for stage base is: 7e72c870614");
@@ -524,13 +584,13 @@ public class BuildMojoTest extends AbstractMojoTest {
         verify(log, times(1)).info("Detected multistage Containerfile...");
 
         // At random verify some lines
-        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base', matches: true");
-        verify(log, times(1)).debug("Processing line: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null', matches: false");
+        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base'");
+        verify(log, times(1)).debug("Processing candidate: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null'");
 
         // Verify stage detection
-        verify(log, times(1)).debug("Found stage in Containerfile: base");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase2");
+        verify(log, times(1)).debug("Processing stage in Containerfile: base");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase2");
 
         // Verify hashes for stages
         verify(log, times(1)).info("Final image for stage base is: 7e72c870614");
@@ -583,13 +643,13 @@ public class BuildMojoTest extends AbstractMojoTest {
         verify(log, times(1)).info("Detected multistage Containerfile...");
 
         // At random verify some lines
-        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base', matches: true");
-        verify(log, times(1)).debug("Processing line: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null', matches: false");
+        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base'");
+        verify(log, times(1)).debug("Processing candidate: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null'");
 
         // Verify stage detection
-        verify(log, times(1)).debug("Found stage in Containerfile: base");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase");
-        verify(log, times(1)).debug("Found stage in Containerfile: phase2");
+        verify(log, times(1)).debug("Processing stage in Containerfile: base");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase2");
 
         // Verify hashes for stages
         verify(log, times(1)).info("Final image for stage base is: 823ed30df4abf7574498d5f766b0ebf70793d73a99fc4220dda200c5131b60ce");